+++ xtask/src/main.rs
@@ -32,30 +32,11 @@ fn try_main() -> Result<()> {
 }
 
 fn lint_all() -> Result<()> {
-    // Lint with only default, only docking, and only freetype
+    // Lint with only default, only docking, and only freetype, and everything
     xshell::cmd!("cargo clippy --workspace --all-targets").run()?;
     xshell::cmd!("cargo clippy --workspace --all-targets --features docking").run()?;
     xshell::cmd!("cargo clippy --workspace --all-targets --features freetype").run()?;
-
-    // Lint winit with all features
-    xshell::cmd!(
-        "cargo clippy --manifest-path imgui-winit-support/Cargo.toml --all-features --all-targets"
-    )
-    .run()?;
-
-    // Lint with various winit versions
-    let winits = &[
-        "winit-19",
-        "winit-20",
-        "winit-22",
-        "winit-23/default",
-        "winit-24/default",
-        "winit-25/default",
-        "winit-26/default",
-    ];
-    for &winit in winits {
-        xshell::cmd!("cargo clippy --manifest-path imgui-winit-support/Cargo.toml --no-default-features --features {winit} --all-targets").run()?;
-    }
+    xshell::cmd!("cargo clippy --workspace --all-targets --all-features").run()?;
 
     // Check formatting
     xshell::cmd!("cargo fmt --all -- --check").run()?;
@@ -71,18 +52,6 @@ fn test_all() -> Result<()> {
     // Test doc examples
     xshell::cmd!("cargo test --workspace --doc").run()?;
 
-    // Test with various winit versions
-    let winits = &[
-        "winit-19",
-        "winit-20",
-        "winit-22",
-        "winit-23/default",
-        "winit-24/default",
-        "winit-25/default",
-    ];
-    for &winit in winits {
-        xshell::cmd!("cargo test --manifest-path imgui-winit-support/Cargo.toml --no-default-features --features {winit} --all-targets").run()?;
-    }
     // Run heavy tests in release mode
     xshell::cmd!("cargo test -p imgui --release -- --ignored").run()?;
     Ok(())